@@ -3,50 +3,49 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@chakra-ui/react": "^1.8.3",
-    "@craco/craco": "^6.4.3",
+    "@chakra-ui/react": "^2.2.4",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
     "@reduxjs/toolkit": "^1.8.1",
+    "@tanstack/react-query": "^4.0.10",
+    "@tanstack/react-query-devtools": "^4.0.10",
     "@types/lodash.debounce": "^4.0.6",
-    "@types/node": "^16.7.13",
-    "@types/react": "^17.0.20",
+    "@types/node": "^18.0.6",
+    "@types/react": "^18.0.15",
     "@types/react-copy-to-clipboard": "^5.0.2",
-    "@types/react-dom": "^17.0.9",
-    "axios": "^0.25.0",
+    "@types/react-dom": "^18.0.6",
+    "axios": "^0.27.2",
     "chart.js": "^3.7.1",
     "clsx": "^1.1.1",
-    "craco-alias": "^3.0.1",
-    "framer-motion": "^5",
+    "framer-motion": "^6.5.1",
     "lodash.debounce": "^4.0.8",
     "moment": "^2.29.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chartjs-2": "^4.0.1",
     "react-copy-to-clipboard": "^5.0.4",
-    "react-dom": "^17.0.2",
-    "react-dropzone": "^12.0.4",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^14.2.2",
     "react-error-boundary": "^3.1.4",
     "react-hook-form": "^7.26.1",
-    "react-query": "^3.34.14",
-    "react-redux": "^7.2.7",
+    "react-redux": "^8.0.2",
     "react-router-dom": "^6.2.1",
-    "react-scripts": "5.0.0",
-    "react-toastify": "^8.2.0",
+    "react-scripts": "5.0.1",
+    "react-toastify": "^9.0.7",
     "redux": "^4.1.2",
     "redux-persist": "^6.0.0",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
   "devDependencies": {
-    "autoprefixer": "^10.4.2",
-    "postcss": "^8.4.6",
-    "tailwindcss": "^3.0.19"
+    "autoprefixer": "^10.4.7",
+    "postcss": "^8.4.14",
+    "tailwindcss": "^3.1.6"
   },
   "scripts": {
-    "start": "craco start",
-    "build": "craco build",
-    "test": "craco test",
-    "eject": "craco eject"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
