@@ -3,7 +3,7 @@ import { Button } from "@chakra-ui/react";
 import { Link } from "react-router-dom";
 import { Input, Checkbox } from "components/Inputs";
 import { useForm, FormProvider } from "react-hook-form";
-import { useMutation } from "react-query";
+import { useMutation } from "@tanstack/react-query";
 
 // redux
 import { useDispatch } from "react-redux";
