@@ -1,4 +1,4 @@
-import { useMutation } from "react-query";
+import { useMutation } from "@tanstack/react-query";
 import { useNavigate } from "react-router-dom";
 import { useForm, FormProvider, useFieldArray } from "react-hook-form";
 import { toast } from "react-toastify";
