@@ -1,162 +1,3 @@
-module.exports = {
-  content: ["./src/**/*.{js,jsx,ts,tsx}"],
-  theme: {
-    extend: {
-      colors: {
-        primary: {
-          // dodgerBlue
-          main: "#605bff",
-          50: "#efefff",
-          100: "#cfceff",
-          200: "#b0adff",
-          300: "#908cff",
-          400: "#706bff",
-          500: "#605bff", // main
-          600: "#4d49cc",
-          700: "#3a3799",
-          800: "#262466",
-          900: "#131233",
-        },
-        main: {
-          primary: "#030229",
-          secondary: "#03022966",
-          disabled: "",
-        },
-        blackAlpha: {
-          50: "rgba(0, 0, 0, 0.04)",
-          100: "rgba(0, 0, 0, 0.06)",
-          200: "rgba(0, 0, 0, 0.08)",
-          300: "rgba(0, 0, 0, 0.16)",
-          400: "rgba(0, 0, 0, 0.24)",
-          500: "rgba(0, 0, 0, 0.36)",
-          600: "rgba(0, 0, 0, 0.48)",
-          700: "rgba(0, 0, 0, 0.64)",
-          800: "rgba(0, 0, 0, 0.80)",
-          900: "rgba(0, 0, 0, 0.92)",
-        },
-        whiteAlpha: {
-          50: "rgba(255, 255, 255, 0.04)",
-          100: "rgba(255, 255, 255, 0.06)",
-          200: "rgba(255, 255, 255, 0.08)",
-          300: "rgba(255, 255, 255, 0.16)",
-          400: "rgba(255, 255, 255, 0.24)",
-          500: "rgba(255, 255, 255, 0.36)",
-          600: "rgba(255, 255, 255, 0.48)",
-          700: "rgba(255, 255, 255, 0.64)",
-          800: "rgba(255, 255, 255, 0.80)",
-          900: "rgba(255, 255, 255, 0.92)",
-        },
-        facebook: {
-          50: "#E8F4F9",
-          100: "#D9DEE9",
-          200: "#B7C2DA",
-          300: "#6482C0",
-          400: "#4267B2",
-          500: "#385898",
-          600: "#314E89",
-          700: "#29487D",
-          800: "#223B67",
-          900: "#1E355B",
-        },
-        linkedin: {
-          50: "#E8F4F9",
-          100: "#CFEDFB",
-          200: "#9BDAF3",
-          300: "#68C7EC",
-          400: "#34B3E4",
-          500: "#00A0DC",
-          600: "#008CC9",
-          700: "#0077B5",
-          800: "#005E93",
-          900: "#004471",
-        },
-        telegram: {
-          50: "#E3F2F9",
-          100: "#C5E4F3",
-          200: "#A2D4EC",
-          300: "#7AC1E4",
-          400: "#47A9DA",
-          500: "#0088CC",
-          600: "#007AB8",
-          700: "#006BA1",
-          800: "#005885",
-          900: "#003F5E",
-        },
-        twitter: {
-          50: "#E5F4FD",
-          100: "#C8E9FB",
-          200: "#A8DCFA",
-          300: "#83CDF7",
-          400: "#57BBF5",
-          500: "#1DA1F2",
-          600: "#1A94DA",
-          700: "#1681BF",
-          800: "#136B9E",
-          900: "#0D4D71",
-        },
-        whatsapp: {
-          50: "#dffeec",
-          100: "#b9f5d0",
-          200: "#90edb3",
-          300: "#65e495",
-          400: "#3cdd78",
-          500: "#22c35e",
-          600: "#179848",
-          700: "#0c6c33",
-          800: "#01421c",
-          900: "#001803",
-        },
-        blue: {
-          50: "#eff4ff",
-          100: "#cedfff",
-          200: "#adc9ff",
-          300: "#8cb3ff",
-          400: "#6b9eff",
-          500: "#5b93ff", // main
-          600: "#4976cc",
-          700: "#375899",
-          800: "#243b66",
-          900: "#121d33",
-        },
-        yellow: {
-          50: "#fff9e9",
-          100: "#ffedbe",
-          200: "#ffe193",
-          300: "#ffd568",
-          400: "#ffc93d",
-          500: "#ffc327", // main
-          600: "#cc9c1f",
-          700: "#997517",
-          800: "#664e10",
-          900: "#332708",
-        },
-        salmon: {
-          50: "#fff4f0",
-          100: "#ffddd3",
-          200: "#ffc7b5",
-          300: "#ffb197",
-          400: "#ff9a7a",
-          500: "#ff8f6b", // main
-          600: "#cc7256",
-          700: "#995640",
-          800: "#66392b",
-          900: "#331d15",
-        },
-        dodgerBlue: {
-          // primary
-          50: "#efefff",
-          100: "#cfceff",
-          200: "#b0adff",
-          300: "#908cff",
-          400: "#706bff",
-          500: "#605bff", // main
-          600: "#4d49cc",
-          700: "#3a3799",
-          800: "#262466",
-          900: "#131233",
-        },
-      },
-    },
-  },
-  plugins: [],
-};
+const tailwindConfig = require("./src/tailwind.config");
+
+module.exports = tailwindConfig;
